Update README.md
My navigation starts with the general information in the banner and shows the location of Hong Kong. I chose light yellow as a background color because Hong Kong is a historical place and light yellow. Since light yellow and light blue are good color combinations, I chose them to be the main colors on the website. Moreover, the biggest challenge I have encountered was the banner design. In order to address this, I studied the lecture notes and we3school to find a solution. 
Soo Tsz Yan 52321945